chore(server): rename fetch entity util
